% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl-v1.R
\name{cluster}
\alias{cluster}
\title{cluster}
\usage{
cluster(data, nr_clusters = c(2:6), interactive = FALSE)
}
\arguments{
\item{var_name}{}
}
\value{

}
\description{
cluster
}
\examples{
imsbasics::clc()
library(tidyverse); library(shp); library(kml)
path <- "data/rawdata/Data SPSS"
data <- open(path) \%>\% longitudinal("PXXF50") \%>\% cluster()
data <- open(path) \%>\% longitudinal("PXXF50") \%>\% cluster(interactive = TRUE)
data <- open(path) \%>\% longitudinal("PXXF50") \%>\% cluster(nr_clusters = 5)
open(path) \%>\% longitudinal("PXXC75") \%>\% cluster()
}
