% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl-v1.R
\name{cross_sectional}
\alias{cross_sectional}
\title{cross_sectional: import a cross-sectional variable}
\usage{
cross_sectional(data, var_name, path. = path)
}
\arguments{
\item{path.}{}
}
\value{

}
\description{
cross_sectional: import a cross-sectional variable
}
\examples{
imsbasics::clc()
library(tidyverse); library(shp); library(kml)
path <- "data/rawdata/Data SPSS"
open(path) \%>\% cross_sectional("AGE04")
open(path) \%>\% cross_sectional("AGE10")
open(path) \%>\% cross_sectional("AGE17")
open(path) \%>\% longitudinal("PXXF50") \%>\% cluster() \%>\% cross_sectional("AGE10")
data <- open(path) \%>\% longitudinal("PXXF50") \%>\% cluster() \%>\% cross_sectional("AGE10")
open(path) \%>\% cross_sectional("P17W39")
}
