% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{data_summary}
\alias{data_summary}
\title{print a brief summary of a dataset (used to have regular information during import)}
\usage{
data_summary(df)
}
\arguments{
\item{df}{A data.frame (needs a columns "id" & "person_haushalt")}
}
\value{

}
\description{
print a brief summary of a dataset (used to have regular information during import)
}
