% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl-v1.R
\name{explain}
\alias{explain}
\title{Title}
\usage{
explain(data, what = cluster, with = as.name(names(data)[length(names(data))]))
}
\arguments{
\item{with}{}
}
\value{

}
\description{
Title
}
\examples{
imsbasics::clc()
library(tidyverse); library(shp); library(kml)
path <- "data/rawdata/Data SPSS"
open(path) \%>\% longitudinal("PXXF50") \%>\% cluster() \%>\% cross_sectional("SEX10") \%>\% explain()
open(path) \%>\% longitudinal("PXXF50") \%>\% cluster() \%>\% cross_sectional("AGE10") \%>\% explain()
data <- open(path) \%>\% longitudinal("PXXF50") \%>\% cluster()
data \%>\% cross_sectional("AGE10") \%>\% explain()
data \%>\% cross_sectional("SEX10") \%>\% explain()
}
