% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_cols}
\alias{import_cols}
\title{import_cols}
\usage{
import_cols(
  file = stop("SHP99_P_USER.sav"),
  path = stop("data/rawdata/Data SPSS/SHP-Data-W1-W20-SPSS/W1_1999"),
  cols = stop(c("IDPERS", "P04W604"))
)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
import_cols
}
\examples{
shp99_p_user_cols23 <- import_cols("SHP99_P_USER.sav", "data/rawdata/Data SPSS/SHP-Data-W1-W20-SPSS/W1_1999", cols = c(2,3))
difficulties to conciliate personal and professional life last 12 months
shp99_p_user_cols_id_p99f09 <- import_cols("SHP99_P_USER.sav", "data/rawdata/Data SPSS/SHP-Data-W1-W20-SPSS/W1_1999", cols = c("IDPERS", "P99F09"))
shp04_p_user_cols_id_p04w604 <- import_cols("SHP04_P_USER.sav", "data/rawdata/Data SPSS/SHP-Data-W1-W20-SPSS/W6_2004", cols = c("IDPERS", "P04W604"))
}
