% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_id}
\alias{import_id}
\title{import_id}
\usage{
import_id(
  file = stop("Please provide a file name in the format xxx.sav"),
  path = stop("Please provide the path to the file.")
)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
import_id
}
\examples{
shp99_p_user_id <- import_id("SHP99_P_USER.sav", "data/rawdata/Data SPSS/SHP-Data-W1-W20-SPSS/W1_1999")
}
