% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl-v1.R
\name{longitudinal}
\alias{longitudinal}
\title{longitudinal: import a longitudinal variable, the longest available sequence will  be taken}
\usage{
longitudinal(data, var_name, path. = path)
}
\arguments{
\item{var_name}{}
}
\value{

}
\description{
longitudinal: import a longitudinal variable, the longest available sequence will  be taken
}
\examples{
imsbasics::clc()
library(tidyverse); library(shp); library(kml)
path <- "data/rawdata/Data SPSS"
open(path) \%>\% longitudinal("PXXF50")
open(path) \%>\% longitudinal("PXXC75")
}
