% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl-v1.R
\name{open}
\alias{open}
\title{open opens and checks a connection to the directory specified by path.}
\usage{
open(path)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
open opens and checks a connection to the directory specified by path.
}
\examples{
open("data/rawdata/Data SPSS")
open("data/rawdata")
open("data/rawdata/Data QWERTZ")
}
