% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbf.R
\name{read_dbf}
\alias{read_dbf}
\alias{read_dbf_meta}
\title{Read .dbf files}
\usage{
read_dbf(path, col_spec = "")

read_dbf_meta(path)
}
\arguments{
\item{path}{A filename of a .dbf file.}

\item{col_spec}{A character vector of length one with
one character for each column. Use "c" to return the
character representation.}
}
\value{
A data.frame
}
\description{
Read .dbf files
}
\examples{
read_dbf(shp_example("mexico/cities.dbf"))
read_dbf_meta(shp_example("mexico/cities.dbf"))

}
