% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shx.R
\name{read_shx}
\alias{read_shx}
\alias{shx_meta}
\title{Query a .shx shapefile index}
\usage{
read_shx(file, indices = NULL)

shx_meta(file)
}
\arguments{
\item{file}{A .shx file}

\item{indices}{A vector of indices or NULL to get all offsets.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{offset} and \code{content_length}.
}
\description{
Query a .shx shapefile index
}
\examples{
shx_meta(shp_example("3dpoints.shp"))
read_shx(shp_example("3dpoints.shp"))

}
