% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp-package.R
\docType{package}
\name{shp-package}
\alias{shp}
\alias{shp-package}
\title{shp: Read Shapefiles}
\description{
Reads shapefiles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/shp}
  \item Report bugs at \url{https://github.com/paleolimbot/shp/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
