% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp-example.R
\name{shp_example}
\alias{shp_example}
\alias{shp_example_all}
\title{Read an example shapefile}
\usage{
shp_example(example_file)

shp_example_all()
}
\arguments{
\item{example_file}{An example filename (see examples for a list of names)}
}
\value{
A full path to the example file
}
\description{
Read an example shapefile
}
\examples{
# all example names
list.files(system.file("shp", package = "shp"), "\\\\.shp", recursive = TRUE)

# get full path of one example file
shp_example("mexico/cities.shp")

# get full path of all example files
shp_example_all()

}
