% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp-meta.R
\name{shp_file_meta}
\alias{shp_file_meta}
\alias{shp_geometry_meta}
\title{Read shapefile meta information}
\usage{
shp_file_meta(path)

shp_geometry_meta(path, indices = NULL)
}
\arguments{
\item{path}{A vector of filenames.}

\item{indices}{A vector of positive integers giving the
indices of the desired features.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Read shapefile meta information
}
\examples{
shp_file_meta(shp_example("mexico/cities.shp"))
shp_geometry_meta(shp_example("mexico/cities.shp"))

}
