% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp-geometry.R
\name{shp_geometry}
\alias{shp_geometry}
\alias{new_shp_geometry}
\title{Create a shapefile geometry vector}
\usage{
shp_geometry(file)

new_shp_geometry(x, file)
}
\arguments{
\item{file}{A vector of filenames.}

\item{x}{A vector of zero-based indices corresponding to the
internal \code{shape_id} within the shapefile.}
}
\value{
A vector
}
\description{
Create a shapefile geometry vector
}
\examples{
shp_geometry(shp_example("3dpoints.shp"))

}
