% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp-aux.R
\name{shp_list_files}
\alias{shp_list_files}
\alias{shp_delete}
\alias{shp_copy}
\alias{shp_move}
\alias{shp_assert}
\alias{shp_extensions}
\title{List files associated with a shapefile}
\usage{
shp_list_files(file, ext = shp_extensions(), exists = TRUE)

shp_delete(file, ext = shp_extensions())

shp_copy(file, to, ext = shp_extensions(), overwrite = FALSE)

shp_move(file, to, ext = shp_extensions(), overwrite = FALSE)

shp_assert(file)

shp_extensions()
}
\arguments{
\item{file}{A .shp file}

\item{ext}{One or more extensions to include}

\item{exists}{Use \code{FALSE} to generate hypothetical shapefile
names.}

\item{to}{Destination: a .shp filename the same length as \code{to} or
a single directory.}

\item{overwrite}{Use \code{TRUE} to}
}
\value{
A list of files of length \code{length(file) * length(ext)}.
}
\description{
List files associated with a shapefile
}
\examples{
shp_assert(shp_example_all())
shp_list_files(shp_example("anno.shp"))

dest <- tempfile()
dir.create(dest)

shp_copy(shp_example("anno.shp"), dest)
list.files(dest)

shp_move(file.path(dest, "anno.shp"), file.path(dest, "anno1.shp"))
list.files(dest)

shp_delete(file.path(dest, "anno1.shp"))
list.files(dest)

unlink(dest, recursive = TRUE)

}
