% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp-meta.R
\name{shp_meta}
\alias{shp_meta}
\title{Read shapefile meta information}
\usage{
shp_meta(path)
}
\arguments{
\item{path}{A vector of filenames.}
}
\value{
A \code{list()}
}
\description{
Read shapefile meta information
}
\examples{
shp_meta(shp_example("mexico/cities.shp"))

}
