% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{yearly_col_names}
\alias{yearly_col_names}
\title{Title}
\usage{
yearly_col_names(cols, year)
}
\arguments{
\item{year}{}
}
\value{

}
\description{
Title
}
\examples{
yearly_col_names(c("IDPERS", "AXXA00"), 2012)
yearly_col_names(c("IDHOUS", "HXXH01"), 2012)
}
