% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{years_available}
\alias{years_available}
\title{Title}
\usage{
years_available(
  file = stop("Please provide a file name in the format xxx.sav"),
  path = stop("Please provide the path to the file."),
  col_name = stop("Please provide a single col_name you want the availability."),
  year_start = "1999",
  year_end = "2017"
)
}
\arguments{
\item{year_end}{}
}
\value{

}
\description{
Title
}
\examples{
years_available("P_USER.sav", "data/rawdata/Data SPSS/SHP-Data-W1-W20-SPSS", "PXXC71")
years_available("P_USER.sav", "data/rawdata/Data SPSS/SHP-Data-W1-W20-SPSS", "PXXC75")
years_available("P_USER.sav", "data/rawdata/Data SPSS/SHP-Data-W1-W20-SPSS", "PXXC106")
years_available("P_USER.sav", "data/rawdata/Data SPSS/SHP-Data-W1-W20-SPSS", "PXXC108")
}
