% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_id}
\alias{import_id}
\title{import_id}
\usage{
import_id(
  file = stop("SHP99_P_USER.sav"),
 
    path = stop("~/FHS/5. Semester/BATHV/dataset_932                                                                     (2)/Data_SPSS/Data_SPSS/SHP-Data-W1-W20-SPSS/")
)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
import_id
}
\examples{
shp99_p_user_id <- import_id("SHP99_P_USER.sav", "data/rawdata/Data SPSS/SHP-Data-W1-W20-SPSS/W1_1999")
}
