% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_long_cols}
\alias{import_long_cols}
\title{import_long_cols}
\usage{
import_long_cols(
  file = stop("P_USER.sav"),
  path = stop("data/rawdata/Data SPSS/SHP-Data-W1-W21-SPSS"),
  cols = c("IDPERS", "PYYW604"),
  year_start = "1999",
  year_end = "2018"
)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
import_long_cols
}
\examples{
library(tidyverse)
shp_p_user_cols_id_p99f09 <- import_long_cols("P_USER.sav", "data/rawdata/Data SPSS/SHP-Data-W1-W21-SPSS", cols = c("IDPERS", "PYYF09"))
shp_p_user_cols_id_p99f09 <- import_long_cols("P_USER.sav", "data/rawdata/Data SPSS/SHP-Data-W1-W21-SPSS", cols = c("IDPERS", "PYYF09"), year_start = "1999", year_end = "2003")
# work conditions stress
shp_p_user_cols_id_w604 <- import_long_cols("P_USER.sav", "data/rawdata/Data SPSS/SHP-Data-W1-W21-SPSS", cols = c("IDPERS", "P04W604"), year_start = "2004", year_end = "2004")
shp_p_user_cols_id_w604 <- import_long_cols("P_USER.sav", "data/rawdata/Data SPSS/SHP-Data-W1-W21-SPSS", cols = c("IDPERS", "P04W604"), year_start = "2004", year_end = "2017")
# interference work - private
}
