% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_SPSS_file}
\alias{import_SPSS_file}
\title{import_SPSS_file}
\usage{
import_SPSS_file(
  file = stop("SHP99_P_USER.sav"),
  path = stop("data/rawdata/Data SPSS/SHP-Data-W1-W21-SPSS/W1_1999")
)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
import_SPSS_file
}
\examples{
shp99_p_user <- import_SPSS_file("SHP99_P_USER.sav", "data/rawdata/Data SPSS/SHP-Data-W1-W21-SPSS/W1_1999")
}
