% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_SPSS_file_head}
\alias{import_SPSS_file_head}
\title{import_SPSS_file_head}
\usage{
import_SPSS_file_head(
  file = stop("SHP99_P_USER.sav"),
 
    path = stop("~/FHS/5. Semester/BATHV/dataset_932 (2)/Data_SPSS/Data_SPSS/SHP-Data-W1-W20-SPSS/W1_1999/SHP99_P_USER.sav")
)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
import_SPSS_file_head
}
\examples{
shp99_h_user_head <- import_SPSS_file_head("SHP99_H_USER.sav", "data/rawdata/Data SPSS/SHP-Data-W1-W20-SPSS/W1_1999")
shp99_p_user_head <- import_SPSS_file_head("SHP99_P_USER.sav", "data/rawdata/Data SPSS/SHP-Data-W1-W20-SPSS/W1_1999")
shp04_p_user_head <- import_SPSS_file_head("SHP04_P_USER.sav", "data/rawdata/Data SPSS/SHP-Data-W1-W20-SPSS/W6_2004")
}
