% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl-v1.R
\name{open}
\alias{open}
\title{opens the specified path. Checks, whether that path actually exists.}
\usage{
open(path)
}
\arguments{
\item{path}{The path, where the SHP data in SPSS format is stored}
}
\value{
a path. Attention: The path is also stored to the global environment.
}
\description{
opens the specified path. Checks, whether that path actually exists.
}
\examples{
open("data/rawdata/Data SPSS")
open("data/rawdata")
open("data/rawdata/Data QWERTZ")
}
