% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbf.R
\name{read_dbf}
\alias{read_dbf}
\alias{dbf_meta}
\alias{dbf_colmeta}
\title{Read .dbf files}
\usage{
read_dbf(file, col_spec = "?", encoding = NA)

dbf_meta(file)

dbf_colmeta(file)
}
\arguments{
\item{file}{A filename of a .dbf file.}

\item{col_spec}{A character vector of length one with
one character for each column or one character to be used
for all columns. The following characters are supported
(designed to align with \code{\link[readr:cols]{readr::cols()}}):
\itemize{
\item "?": Use DBF-specified field type
\item "-": Skip column
\item "c": Character
\item "i": Parse integer
\item "d": Parse double
\item "l": Parse as logical
}}

\item{encoding}{Use \code{NA} to automatically guess encoding,
\code{""} to use system encoding, or a length-one character
vector overriding the automatically detected encoding.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Read .dbf files
}
\examples{
read_dbf(shp_example("mexico/cities.dbf"))
dbf_meta(shp_example("mexico/cities.dbf"))
dbf_colmeta(shp_example("mexico/cities.dbf"))

}
