% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp-meta.R
\name{shp_meta}
\alias{shp_meta}
\alias{shp_geometry_meta}
\title{Read shapefile meta information}
\usage{
shp_meta(file)

shp_geometry_meta(file, indices = NULL)
}
\arguments{
\item{file}{A vector of filenames.}

\item{indices}{A vector of positive integers giving the
indices of the desired features.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Read shapefile meta information
}
\examples{
shp_meta(shp_example("mexico/cities.shp"))
shp_geometry_meta(shp_example("mexico/cities.shp"))

}
